read new config.schemas.module property
