+++ test.js
@@ -1,0 +1,3 @@
+var schemas = require('./index');
+
+console.log(JSON.stringify(schemas, null, 2));
